+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+## Added
+* new option `authSpecOptions`.  This will get added verbatim to the result of authenticationSpecification function
+
 ## [1.0.0] - 2018-05-22
 
 ### Added